B8 add authentication

+++ controllers/admin.js
@@ -5,7 +5,6 @@ exports.getAddProduct = (req, res, next) => {
         pageTitle: "Add Product",
         path: "/admin/add-product",
         editing: false,
-        isAuthenticated: req.session.isLoggedIn,
     });
 };
 
@@ -50,7 +49,6 @@ exports.getEditProduct = (req, res, next) => {
                 path: "/admin/edit-product",
                 editing: editMode,
                 product: product,
-                isAuthenticated: req.session.isLoggedIn,
             });
         })
         .catch((err) => console.log(err));
@@ -66,22 +64,24 @@ exports.postEditProduct = (req, res, next) => {
 
     Product.findById(prodId)
         .then((product) => {
+            if (product.userId.toString() !== req.user._id.toString()) {
+                return res.redirect("/");
+            }
             product.title = updatedTitle;
             product.price = updatedPrice;
             product.description = updatedDesc;
             product.imageUrl = updatedImageUrl;
-            return product.save();
-        })
-        .then((result) => {
-            console.log("UPDATED PRODUCT!");
-            res.redirect("/admin/products");
+            return product.save().then((result) => {
+                console.log("UPDATED PRODUCT!");
+                res.redirect("/admin/products");
+            });
         })
         .catch((err) => console.log(err));
 };
 
 // GET /products
 exports.getProducts = (req, res, next) => {
-    Product.find()
+    Product.find({ userId: req.user._id })
         // .select('title price -_id')
         // .populate('userId', 'name')
         .then((products) => {
@@ -89,7 +89,6 @@ exports.getProducts = (req, res, next) => {
                 prods: products,
                 pageTitle: "Admin Products",
                 path: "/admin/products",
-                isAuthenticated: req.session.isLoggedIn,
             });
         })
         .catch((err) => console.log(err));
@@ -98,7 +97,7 @@ exports.getProducts = (req, res, next) => {
 // POST /delete-product
 exports.postDeleteProduct = (req, res, next) => {
     const prodId = req.body.productId;
-    Product.findByIdAndRemove(prodId)
+    Product.findByIdAndRemove({ _id: prodId, userId: req.user._id })
         .then(() => {
             console.log("DESTROYED PRODUCT");
             res.redirect("/admin/products");
